initial commit of vala+gdesktopmenu branch
This is a rewrite of alacarte using vala and other modern
technologies. The main goal is to create a cleaner and simpler
code base that solves several design probems with alacarte.
Changes made to menus using the old version of alacarte will
mostly still work in this version but the plan is to no
longer show NoDisplay items in alacarte as these tend to
confuse users. Through xdg-list I am currently working out
a better way to handle hiding things specifically for or in
a menu editor like alacarte.

Oh, and this version of alacarte depends on libxfce4menu.
This clean and fast library is actively maintained and I am
working with the main developer to make it more suitable for
use in GNOME applications. We will soon be renaming it
gdesktopmenu to reflect this goal. Currently it depends on
the parser-rewrite branch [1] but hopefully this will be
merged to master soon or moved to a completely new location
once the rename is finished.

[1] http://git.xfce.org/jannis/libxfce4menu/log/?h=parser-rewrite