Fix crash in on_menu_tree_cursor_changed()
This bug is caused by the fact that
treeview.get_selection().get_selected() can return a None iter when the
selection has no rows (e.g. if the user has deselected a row by
Ctrl-clicking on it).

The solution is to use gtk.SELECTION_BROWSE selection mode; it will
ensure that one row is always selected and valid iters are returned from
the selection's get_selected().

https://bugzilla.gnome.org/show_bug.cgi?id=652454