MainWindow: Use html.escape.
Python 3.2 deprecated and 3.8 removes the escape function from the cgi module.
The official documentation states to use html.escape as replacement.

This makes Alacarte work with Python 3 for me.

As @muktupavels pointed out the default parameter for the quote option
has changed when moving from cgi to html. This patch sets quote to
False, keeping the parameter name for better readibility.

See:
https://docs.python.org/3.7/library/cgi.html#cgi.escape
https://docs.python.org/3.8/library/html.html#html.escape